[Bug 354]:  The case of this bug related to /agent/volatile/arp is fixed, but I'm not sure that it works for the case of more depth.
Let's describe the problem in more details (note that there was no obvious
description of the problem in bug posting):

we have /agent/volatile/arp subtree, where dynamic ARP entries reside,
then we can run synchronize operation and it leads to:
1. "list" operation on the Agent (it returns all the IP addresses reside in
   dynamic ARP table);
2. "get" operation on the Agent (it returns MAC address for the IP address);
   But so far we did not get and locks after the first operation we can't be
   sure in general that ARP entry is not died just after "list" operation
   completes and "get" has not started yet.
   In this case "get" operation returns NOSUCHINSTANCE error, as if we
   passed something non-existing.

   To bypass the problem we should care that Agent returns this particular
   error for volatile subtree, and then just ignore it, as if there was no
   ARP entry for the particular IP address.

What about bug fixing? Is it really fix all the cases?
I'm not sure that this patch will fix the situation when we have subtree of
the depth more than two, for example:
/agent/volatile/XXX/YYYY/ZZZZ;

In that case we can fail on getting "/agent/volatile/XXX/YYYY/ZZZZ" object,
and delete "/agent/volatile/XXX/YYYY/" subtree, but we have to remove
""/agent/volatile/XXX/".

So I'm not sure that this is a real fix.



git-svn-id: https://svn.oktetlabs.ru/svnroot/prj/oktetlabs/test-environment/src/trunk@13398 9d14d920-75d5-0310-8bd7-a52572981aee