Check msg_flags in case of success and in case of failure
rpc_recvmsg() and rpc_recvmmsg() were already checking that msg_flags are
set to 0 in case of success.  This patch checks that msg_flags fiels is not
chenged in case of failure.

Also, this patch changes TEST_VERDICT to RING_VERDICT: I do not see any
reason to stop the test if it spoils msg_flags.


git-svn-id: https://svn.oktetlabs.ru/svnroot/prj/oktetlabs/test-environment/src/trunk@92737 9d14d920-75d5-0310-8bd7-a52572981aee