Bug 8746: Provide a way to build unix TA with DPDK
Include DPDK libraries in the symbol table (used for RPC as well).

It cannot be done via builder.conf since -Wl,-whole-archive is required
to pull constructors. May be it is not required in fact sync all objects
are pulled because of symbol table, but it may change in the future.
Also it -Wl,-whole-archive is a way how DPDK link to apps in its
makefiles.