Bug 8761: remove manual handling of function getpid
Previous patch begins to check the value of the variable TARPC_DL_NAME.
And now it is a useless code.

During the creation of RPC server the RPC function getpid is called without
specified library name. And later on agent side, function tarpc_find_func
calls the function tarpc_setlibname with the empty library name. Then attempt
to call this function with the specified library name leads to the error:
  "Dynamic library has already been set to ..."