cleanup: fix warning
In general the prototype of the function is wrong. If it gets const
value, it cannot cast the const qualifiers from its derivatives.
But it does so and it is spread in the code. There is no easy way to
fix it.

It is absolutely safe to drop 'const' qualifier to pass variable to out
only parameters, since it simply means that it is used more stricly in
the caller function.