Bug 9248: extend RCF confstr to support new token 'connect='
Allow specify IP address or hostname where RCF should connect to reach the
test agent. It can differ from the original 'host' where test agent is copied
to, for example if the agent is run in non-default network namespace.
IP from the auxiliary namespace cannot be used as host because the namespace
may not have sshd, so test agent cannot be copied there using scp. But it can
be copied to 'host' and run in the net namespace, then RCF can connect to the
test agent using IP address which is granted to an interface in the namespace.