Bug 9384: add delay to wait until iperf server has started
In some cases especially on slow machines it is possible the server
actually starts later than client. We need to have some guarantee
the server has started (and is listening the port) by the time a user
starts the client. Since we cannot determine such moment exactly,
the simple delay is presented here.