Bug 9504: add conversion of ASN.1 boolean to te_bool at the asn_read_bool()
Since te_bool is bool of <stdbool.h> now we should use it properly. I.e. we
cannot interpret it as interger value since allowed values for it are {0, 1}
only.
But ASN.1 interprets it as integer value {0, 0xff}. Here is a problem in
translating ASN boolean value to bool, because we get it by pointer of int
and compilator cannot detect that bool value is set to unexpected value
(different from 0 and 1).
Experiment shows that gcc operates the bool value as bit-field and performs
some optimization when ternary operator is used.
The following example demonstrates the issue:
    te_bool asn_bool_val;

    asn_read_bool(asn_val, &asn_bool_val, "");
    WARN("asn_bool: %d, !asn_bool: %d, te_bool: %d, !te_bool: %d",
         asn_bool_val, !asn_bool_val,
         asn_bool_val ? TRUE : FALSE,
         asn_bool_val ? FALSE : TRUE);

Output:
    asn_bool: 255, !asn_bool: 254, te_bool: 255, !te_bool: 0

To avoid this we have to convert ASN boolean value to bool explicitly.