Bug 8751: Fix handling of RSS HF combinations in terms of TE
Currently TE maps RTE RSS HF combinations to single-bit flags
which prevent tests from checking individual HF bits. This is
outrageous and must be fixed.

This change will crush log readability but this is fair. This
must be resolved as a separate patch which could replace
the corresponding macros with functions to catch multiple
bits forming short combinations and log their names.

Fixes: 3f9a21b6d675 ("Bug 8751: Mapping of the RPC RSS hash protocols in RTE and vice versa")