Bug 9659: iperf "read failed: Connection refused" is not an error
iperf server can print the mentioned message in its stderr while stdout contains valid data.
Also it can be printed in case of broken stdout data, that is why we should look for [SUM]
records only if iperf runs several streams: the only way to detect broken report because it's
valid if is doesn't contain [SUM] in case of 1 stream
Since now it's necessary to pass number of stream to tapi_perf_server instance!