Bug 9491: ignore host keys validation when specifying ssh key
This is a lazy code. We use -i in several cases, but one of the most
common is vhost that might change it's identity from boot to boot. And
people might redeploy it w/o telling.

Proper implementation should add another option into RCF conf file,
but for now this one is good enough.