Bug 9666: we should make pipe buffer big enough so that one can fit the entire FIO report
The issue we're trying to solve is that pipe buffer is full, FIO
process ends and tries to write JSON output into the buffer. But it
can't cause it's full. And waitpid hangs cause the process is still
alive. One alternative would be to read output from time to time, but:

 o this is hardly perfect, cause simple reads (with the current APIs)
   will fail and polute log with errors,
 o periodic reads are not supported on the engine (they should be),
 o and I want something fast and simple.

So the simplest solution is to increase default pipe buffer size. It's
not done on the "fast path", so no reason to be worried by the
performance.