Bug 9695: implement TST PCO type and rename 'tester' into 'unspec' while keeping things backwards-compat.
### 1. Motivation/problem

Current design is very 'sockapi'-oriented. Example:

'net':IUT
{
  'hostA' {
    { 'my_pco' : tester},
  ...
  }
}

while having

    /net:mynet/node:A = 1
    /net:mynet/node:B = 0

does not mean that I'm getting a non-IUT host, cause tester just means that we're not planning to use the IUT capabilities.

So we have:

  - slightly non-obvious logic,
  - no support for strict hosts types, i.e. if I have 4 hosts - 2xEachType
    and I want to say (in env) that I need a back to back connection between
    them -- can't be done

Current system can't satisfy all the requirements that we're now facing.

### 2. Short term solution

I'm adding a hack, we'll have 4 types:

  - IUT : only IUT,
  - TST : only TST,
  - unspec : current 'tester'
  - tester : mapped to 'unspec' cause I don't want to touch any of the
    trc/package.xml