Bug 9205: use /bin/bash instead of SHELL that might be missing on the target host
If we're building TE under non-bash shell (which should be supported!)
we will have SHELL set to something like zsh. Then on the target it
would be missing.

Why not /bin/sh -- we're sourcing te_functions on the target and file
is said to be written in bash. Thus bash should be present. It's a
limitation, but it's better then fighting with bashisms silently
breaking things.