Bug 8744: Restructure and generalise EAL reuse mechanism
The current code gets the number of RTE devices and then
performs multiple actions in a loop to reset their state.
However, this approach might become unacceptable as soon
as support for bonding and failsafe ports is implemented.

Slave devices maintained by such ports must not be reset
independently until released. Accessing RTE ports one by
one does not indicate the difference between masters and
slaves. A better approach is to look through the list of
network interfaces maintained by the environment so that
it is possible to take their resource types into account.