Bug 9695: add IUT_PEER net node type and PCO type
### Why

The primary usecase I'm trying to solve is be able to say that we have
NVMeoF initiator and target and they should go to certain nodes in the
network topology. It happens so that IUT/tester are not enough, cause
they're talking in terms of "we're pre-loading a certain library to
make the PCO an IUT one".

While what we sometimes need is to say that nodes have certain types
and when binding the PCOs those types should be honored.

This is a second attempt to do this. Previous one e846adebd232 was
rolled back to make sure it's hard to read file annotation.

### Proper solution

It's known that types of 0/1/2 are rubbish and hard to read. So a
better solution will be implemented, but later. Sorry.