Bug 9725: add tapi_sh_env_ta_path_append() to append directory to PATH of the TA
Right now we have /local:ta/env:PATH, but it will remove local TA
directory from PATH. And most of the APIs in tapi_sh_env take PCO
which is not super-convenient in case of prologue. We do not need
to restart PCOs, so it's fine to give just the TA name.