Bug 9578: add sync of MTU value to tapi_set_if_mtu_smart_aux()
If there are eth1 interface and vlan interface eth1.5, and we
change eth1 MTU from 1500 to 500, then eth1.5 MTU is changed
automatically by OS.  When we restore eth1 MTU, then eth1.5 is
not restored by OS, but Configurator says that eth1.5 has 1500
and tapi_set_if_mtu_smart_aux() doesn't restore value.
So we have to update Configurator values and ensure that all MTUs
are restored.