Bug 8744: Exercise YAML configuration file processing in Configurator
This change is needed by DPDK testing in order to have more elaborate
way of instance creation in the project-specific tree. The idea is to
add a limited set of instances in accordance with logical expressions
specified in a YAML document either for individual instances or for a
group of entries. The new facility will produce an XML document to be
consumed by the primary configuration file during Configurator launch.

For the sake of simplicity, only ordinary expressions are implemented
to compare environment variables with constant values. The statements
may be specified in a sequence within a single conditional node which
provides AND behaviour for the given statements. Also, one may deploy
multiple conditional nodes (statement groups) to achieve OR behaviour.

A brief example is as follows (comments start with # sign):

add: # <-- dynamic history command name
  # one-line conditional node applicable to all instances:
  cond: TE_ENV_ALWAYS_TRUE == true
  instances: # <-- command property containing a list of instances
    - "/local:/dpdk:/offloads:" # <-- instance entry
    - "/local:/dpdk:/offloads:/dev:"
    - "/local:/dpdk:/offloads:/dev:/tx:"

add:
  instances:
    - "/local:/dpdk:/offloads:/dev:/tx:/fixed:MBUF_FAST_FREE":
      cond: # <-- conditional node for the given instance
        # The list of statements in the node (AND behaviour)
        - TE_ENV_RTE_VERSION >= 18080000
        - TE_ENV_SFC_DP_TX == ef10_simple
      # Alternative check (or a group of checks) to be evaluated when
      # the previous check results in FALSE (OR behaviour):
      cond: TE_ENV_SOME_VARIABLE == whatever

One has to follow indentation rules, as shown in the example, and use
double quotation marks with instance names because the latter include
colons which also denote YAML node mapping and must not dangle unless
there is a real intention to indicate some child construct afterwards.

The code is supplied with conditional compilation to escape breakages.
Support for the new facility must be enabled beforehand by the option:

TE_APP_PARMS([configurator], [--with-conf-yaml], [], [], [], [])

Also, LibYAML and its development package must be present on the host.