Bug 9736: improve TA objects API
Implement ta_obj_find_create() which creates an object if it cannot
find it.
Improve ta_obj_value_set() so that it can fill object's fields
if a new object should be created. Setting a value may be a part
of commit which changes existing system object. In that case we need
to create corresponding TA object and fill its fields with existing
values, so that we can edit them and so that after commit fields
which we do not touch will be left unchanged.