Bug 9847: ignore session name in TRC report
Session name is ignored because currently Tester ignores name
attribute of <session> tag in package.xml, instead using name
attribute of parent <run> tag. And it does not match this name
to TRC database, but only allows to use it when specifying test
path to be executed.
For consistency the same should be done here if session name is
printed to RAW log.