Bug 9738: add macros to log artifacts on different log levels
This is needed to be able to do auto-processing of the results and
auto-classification of the meta information supplied by the test. For
instance, if the test sees a failure and starts to collect important
data that is related to the root cause of the failure - this sounds
like an ERROR artifact. While if the test just logs performance number
- it's a simple RING one.