Bug 9875: use asn_get_subtype() to find type of the .type object we're expanding
Old code was doing top down search taking each token from a.b.c.[array
is here] and finding corresponding type. Problem was that if we need
a.b.c type it was failing on 'a.b' - it found it and it's not a
SEQUNCE_OF as the code was expecting.

New logic takes the [ from the tail of the path and finds ASN type
that describes the array. And then it takes subtype cause we're not
expanding the array, but rather an array element.