Bug 9941: return NULL from te_sockaddr_get_netaddr() in case of NULL addr
It can be used when creating CSAP inside a function to which sockaddr
arguments are passed. With this change we can simply pass such arguments
to te_sockaddr_get_netaddr() when creating a CSAP, so that if some
sockaddr pointer is NULL, NULL will be passed to CSAP creation function
(meaning that related address is not important).