Bug 9846: rcf_pch: errors to warnings when grabbing locked resource
A process can learn if a resource is locked by another process only by
trying to grab the resource. If the resource is locked, RCF PCH will
spam errors into log. But it is not a critical failure in case of multiple
resources being present (e.g. the first CPU core is locked, but other
cores are available).