Bug 9567: allow not to run unnecessary prologues/epilogues in Tester
Currently, if test iterations are specified not (only) via test path
(including parameter values) but (also) via requirements, unnecessary
prologues/epilogues may be run for test sessions in which not a single
iteration matches requirements from command line. Also this may happen
if a hash is used to specify iteration to run, and test iterations are
inside session which may be iterated with different values of parameters
as a whole. Then prologues/epilogues are run for each iteration of this
session, even if only one of the session iterations includes the test
iteration to be run. This patch fixes these issues by introducing
preparatory testing configuration walk during which improved scenario
is created including only those iterations which are actually going to
be run according to hashes/requirements.