Bug 9911: tapi_cache: Initial implementation
1 Implementation of cache area function to register actualization callbacks

2 Implementation of functions to actualize and invalidate cache areas
Requirements:
2.1 Invalidating function should not remove instances of objects with registered
callbacks. It could be useful to keep such instances alive since they look
like permanent stuff, and to satisfy a requirement about possibility of
actualizing a cache with walking throughout the instances with registered
methods
2.2 It should be possible to actualize the full cache, or any other top-level area
which does not have callback but whose children have ones. In other words if it
fails to find callback for requested instance of area it should try to find it
at children. It should stop searching for callbacks as soon as it is found. It
should invoke callback for all brothers of the instance.

3. Implementation of functions performing direct actions on cache area

It has been tested with ${TE_BASE}/suites/tdd_tapi_cache, which is intended to
verify TAPI cache functionality