bug9302: do mkdir .libs with -p flag to avoid warnings
I'm not sure the change is not hiding a problem, ie. rm -f .libs can
keep .libs if it was not empty, so perhaps rm -rf .libs is a better
fix. But I don't think it matters and mine is safer.

Before this change I observed (ubuntu 18.04) errors saying that it
can't create a directory that already exists.