Bug 9973: fix TRC to use the first match result
Use a singly-linked tail queue which supports insertion to tail
easliy instead of a singly-linked list.
It makes it easy to keep results order (it was stored in reverse
order) and makes matching to pick the first match.
To keep functionality all insertions to head should be substituted
with insertions to tail.