cleanup: set close on exec flag on descriptors opened by te_exec_child()
te_exec_child() opens pipes to communicate with a child process. The
file descriptors remain open after the call and because close-on-exec
flag is not set by the function, these desctiptors will not be closed
after exec of another process (for example, if te_exec_child() is used)
several times.

Set the close-on-exec flag to avoid descriptor leaks.