Bug 9859: add a timeout to job_destroy()
In case when job_destroy() is called before a process termination,
the process must be killed. The default behaviour is to kill it with
SIGTERM, though the process might intercept the signal, so additional
time might be needed for wait(2) to free the process. If the wait
times out, the best option is to kill the process with SIGKILL.

Allow to set a specific timeout in job_destroy() to wait before
kill() with SIGKILL is called.