Bug 10057: add functions to add/del a simple route to destination address
It is quite inconvenient to operate sin(6)_addr in the tests, much better
to delegate its extraction from sockaddr to a lib function and operate
only sockaddr in the test independently on IP version. Also it is worth
to have a similar way to add a route without a gateway.