Bug 8744: TAD: Don't insert empty IPv4 options DU to a raw ASN.1 packet
If an ASN.1 raw packet containing IPv4 PDU is converted to a
pattern, it will be impossible to actually start receive
operation because IPv4 PDU will contain empty IPv4 options.
Tests which need to make patterns from ASN.1 raw packets
have to traverse the resulting patterns and remove the
faulty field from each IPv4 PDU encountered. This is
humiliating and inefficient.

Workaround this on Rx post-match step to avoid adding
empty IPv4 options DU to IPv4 PDU on the step.