Bug 8744: tapi_dpdk: get port number from EAL arguments
When "--vdev" argument is used to initialize EAL (to setup failsafe
or bonding), port number of virtual device is not 0 in case of
also having physical devices whitelisted.

The solution is not ideal and rely on EAL behaviour, but we simply
have no better ideas which could be implemented fast enough.