Bug 8977: allocate port if port in address is not initialized yet
Usually port is allocated in TEST_GET_ADDR(), but it is not used
in the case of DPDK testing since addresses are not directly used
in the test body itself. It is just used in environment substitution.

It is important to save allocated port in the address to be sure
that the same value is obtained the next time if referenced many
times.