Bug 10294: add a possibility to reserve space for string
In most cases stepwise reallocation is not handy, in some cases
it is possible to predict the space required. This patch adds
a possibility to reserve space, and changes the mechanism extra
length is chosen. First, we check if grow factor of 2 is fine,
if not, we check 4, 8, 16 (bot no more), until we find a
satisfying size. If not found still, simply add requested length
to the buffer. Second, the malloc overhead circumvention is added
for uniformity with standard C++ library (std::string)