Bug 10218: rcfunix: support SSH proxy to connect to TA
Virtual machines use QEMU port forwarding to provide SSH and RCF
connections, however, these ports are accessible from hypervisor
host only. Use hypervisor host as an SSH proxy.

It is assumed that RCF port is not directly accessible in this case as
well and should be forwarded by SSH. Right now try to use exactly the
same port locally for simplicity, but it can be choosen dynamically
in the future.