Bug 10332: loggerta: fix crash on NULL log user
When log user is passed via the first argument, the string data
is written into ring buffer, and string pointer into ring buffer
header. It is possible allocate and fill header successfully and
fail allocation of space needed for string data in ring buffer. This
partial initialization of log message in ring buffer causes crash.

Make log funtions restore previous state of ring buffer in case of
partial initialization of log message. Such a message is dropped
and that fact is indicated by a gap in log sequence numbers (which
is reported in Logger). This will ensure that only messages with
all arguments initialized pass.