Bug 10619: make job wait idempotent
It is possible to store the exit status of a job
after it has been waited for. Modify the job control facility
to allow job wait to return the status of the last finished process,
when no process is running. Though an error will be returned if
there were no started processes.