Bug 9682: enforce static library to be built for TA side
It is already done on platform level, but if TA is built for default
platform, it builds shared library if shared libraries are enabled.

Static libraries should be linked to TA to avoid necessity to copy
shared libraries to remote host and explain linker where to find them.

Also some libraries depend on symbols defined in TA.
It is not a problem in the case of static library build.