Bug 9051: implement sockaddr_h2str_buf()
sockaddr_h2str() is often used multiple times for the same format
string - for example, when both local and remote addresses should
be printed out. The problem is that this does not work for
TE-specific address family TE_AF_TARPC_SA. This patch makes
function more predictable, so that it always uses the single
buffer and should not be expected to work correct when used
multiple times for the same format string; sockaddr_h2str_buf()
should be used instead in such case.