Bug 10666: remove TSA_TST_USE_LINGER from tapi_tcp_states lib
It was implemented incorrectly: zero linger was set for both IUT
and Tester sockets which resulted in unintended RST from IUT
in a test which called close() on IUT socket.
It seems more convenient always to set SO_LINGER with zero
linger value on Tester socket, so that RST packet to IUT can
easily be generated by closing it if requested. There is no need
in a special flag asking to set SO_LINGER on Tester.