Bug 10578: te_mi_log: make tool name a mandatory field in MI log
The information about the tool that did statistics gathering is
crucial and should be always logged along with statistics since
the information may not be present in test name or parameters.
Also the tool version can be added as suffix to the name if there
are significant differences in versions.