Bug 9682: install changed files only in the case of suite meson build
Use meson install --changed-only which is introduced in version 0.47.0
which is OK since it is required on Test Engine hosts only.

It detects files removed from installation path, but does not detect
other changes (e.g. file removal and creation of empty file with the
same name). I.e. installation is triggered by either missing target
file or changes in the built file to be installed.