Bug 10130: move RPC server SID storage to RPC server on TA side
/rpcserver_sid objects are used to store RCF session ID allocated for
the RPC server. Storage is required to avoid session creation each
time when RPC server is obtained in test.

The problem comes when test agent is destroyed and started again.
As the result, invalid SID (sessions are cleared) is preserved in
the storage and the first RPC request to the server fails on RCF
level since the session does not exist.

New configuration tree stores the SID inside RPC server subtree on
TA side and therefore automatically cleared if TA is destroyed and
started again.

From RCF point of view a new behaviour should be similar to the old
one in the case of reused RPC server. If RPC server is created/destroyed
for each test (not reused), new session will be created and I think
it is good, since it a new RPC server each time.

Storage of SID on TA side requires a bit more efforts to obtain it,
but it should not be a problem since it happens on RPC server get only
(not per RPC call).