Bug 7096: use pkg-config to find libtirpc
pkg-config is the right way to find libraries and get required
CFLAGS and libraries to link with.

Unfortunately, autotools-based build does not have a mechanism
to inherit dependencies (CFLAGS, libraries) and it is required
to do it in every library which includes corresponding headers
even indirectly. Hopefully, autotools-based build will die soon.