Bug 10784: replace non-standard TE_STRNCPY by TE_STRLCPY
There are two issues with TE_STRNCPY

1. The function arguments order differs from standard strncpy

2. Forced terminating truncated too long string is closer rather to strlcpy
than to strncpy. Since we already have te_strlcpy (Bug 10699), it's worth
renaming it to TE_STRLCPY

Since the function is intended to detect a string truncation and log
the error, it is useful to return a pointer to destination string instead of
either status code or size of source string (as in case of strlcpy).
For example, it allows to pass the function return value as argument of some
other function