Bug 8744: cm/local_dpdk: add node to control extra Tx descriptors required
Some tests calculate required number of Tx descriptors and use it as the
Tx ring size. Typically, it is just a number of segments in a packet.
However, some HW/SW requires extra descriptors (e.g. virtio for net
header). Provide a way to specify it in local configuration tree.
It is not required to make it a precise value since algorithm could be
really complicated and too HW specific, just estimated maximum number
of extra Tx descriptors which could be required.