Bug 10879: agents/unix: forget VM PID once QEMU terminates
The current version of vm_is_running simply makes a call to ta_waitpid with
the same cached PID value. While this is probably good enough in practice,
in theory this can lead to a situation where this function mistakes some new
process with the same PID for the original QEMU process.

In this patch I clear out the cached PID based on the return value of
ta_waitpid.