Bug 10879: lib/confapi: add method to get CS value size
In some cases, e.g. checking if a buffer is large enough, it is useful to
know the size of a given value. Currently the only way to do so seems to be
to determine the type of this value and then manually determine its size,
for example, by calling the strlen function.

This leads to duplication of code and, when it comes to address values,
quite significant amounts of code.

In order to provide a simple uniform way to achieve the same result, this
patch introduces a value_size function as a member of cfg_primary type.