Bug 10879: lib/confapi: add a library to make API messages
Whenever Configurator needs to use its own API, it needs to construct the
corresponding messages. The current approach is problematic: it leads to
code duplication (since these messages have to be constructed in conf_api
anyway), and inflation (constructing a message takes quite a bit of lines).

In order to improve the situation, it was proposed to create a separate
library whose sole responsibility would be constructing these messages. This
library can then be used in both the conf_api library and in Configurator
itself.

This patch marks the start of work on this new library: it adds the necessary
files and defines the function to construct CFG_GET messages, which is now
used in conf_api.