[1/2] Bug 10939: remove tools/rgt/rgt-raw2xml
This implementation was crafted as "improved" version of RGT that
should not be memory-hungry and should simplify development.
Yet the result is significantly slower than C implementation (may be
Lua is slow for the task, may be it is cooked bad), implementation is hardly
perfect and incomplete, and the author gave up. Thus it makes no sense to keep
it in the tree "just in case". Especially given it generates piles of warnings
during documentation creation.