Bug 8714: extend DPDK device arguments local configurator tree
Add vendor and device IDs to the name of device argument tree entry.
It makes it possible to specify devices with which vendors/device IDs
should get which device arguments. Also, device arguments common to
all devices or all devices with specific vendor can be specified.

Add a TAPI to get device arguments of a PCI device.