Bug 8714: rpcc/dpdk: add hotplug wrappers that reset EAL arguments cache
When reuse PCO parameter is enabled, RPC server is not restarted on
each test iteration as long as required EAL configuration stays the
same. The hotplug calls change that configuration, but the RPC
functions does not reset the cached configuration string.

Add wrapper for RPC functions that take care of that.