Bug 10965: lib/log_proc: add entity-user filters
This is a reimplementation of the code from tools/rgt/rgt-cfg.

This module lets the user create and use message filters in the form of C
structures instead of Tcl code, which is simpler and will allow us to remove
dependency on Tcl.

Moving this functionality into a separate library will allow other TE
entities to use it, which will help to resolve other issues.