Bug 10139: conf/vm: support vhost-user interface type
The interface type allows VM to connect to OVS.
The connection is established via Unix domain socket which
should be specified as a character device. It also should
be in a server mode.

Also for that interface to function, memory should be backed
by hugepages file and the memory should be assigned to a NUMA
node. These additional agruments are added automatically
and cannot be configured yet.