Bug 10936: tools/rgt: delay JSON unpacking until all fields are known
Since the event type is now determined by MI type, we can unpack the JSON
object according to the specific schema of this event. This means that we
now know all possible fields at the time of unpacking, which lets us enable
strict unpacking.

In addition to being cleaner, this approach can detect typos in field names.