Bug 8744: tapi_tad: modify VLAN tag APIs
The APIs do not handle missing VLAN tag fields specification
properly. The fields can be either set to plain value, specified
in terms of a range (with a mask) or not set at all. The users
of these APIs do not need range specification, but require that
unset fields are handled correctly.

Update the APIs to differentiate between zero and unset values
and specify VLAN tag fields separately for vlan injection
to set only required fields.