Bug 9205: execute SSH commands with /tmp as current working directory
Cannot cd to caller $PWD, since it could be in a home directory not
accesible on remote host or, even more, located in local directory
which is not exported over NFS at all.

Provide a function with old behaviour and use it autotools-based
build which relies on NFS.

Do the same in non-SSH case as well to be consistent.