Bug 11029: tools/rgt: stop treating the status field as optional
The presence of the status field is no longer used to distinguish between
test start and test end messages. Now we always know which fields to expect
during unpacking, and the status field is always present in test_end
messages.

This patch reflects this with a change in the json_unpack format string.