Bug 11029: suites/selftest: add tests that yield artifacts and verdicts
Right now there are no artifacts or verdicts in the log if the run is
successful. This makes debugging of TRC and RGT functionality more
complicated.

Adding these tests will also give potential new developers on this project
an easy way to see what artifacts and verdicts are and how they look in the
code and in the logs, along with an example of a non-trivial TRC config.