Bug 10936: doc/drafts: add schema for MI Tester Control messages
This schema describes the currentily produced messages. A lot of "null" types
can be removed once libjansson is upgraded to at least 2.11 on all setups.
For example, the newest available version on CentOS 7 is 2.10.

The schema would definitely benefit from being split into multiple files, but
JSON Schema relies on absolute URIs for that.