Bug 11092: separate resource add and set
If a resource is added on configurator 'add' operation, the
properties of the resource cannot be configured beforehand.
Separate add and set operations to be able to configure a
resource before grabbing and keep compatibility with old
resource acquisition (if resource is added with value, it is
grabbed right away).