Bug 11088: agentlib: add API to allocate available TCP UDP ports
The port allocation is already present in test API, but
it can't be used in TA since it is bound to RPC server
and allocates ports from a range local to each TE instance.

Add an API to agentlib so that a TA can allocate a port
for an internal use. Later the port allocation will be
also available from configurator.