Bug 11088: add fallback to IPv4 in TCP UDP port allocation
Some systems may not support IPv6 so using IPv6 as a default
protocol for checking whether a port is free is not convenient.
Add 0 socket family to the API that specifies that IPv6 should
be used to check a port, but a fallback to IPv4 might be used
if the IPv6 protocol is not supported.