Bug 11092: grab kernel module as a resource
Add resource acquisition into the module TAPI. The module
is first grabbed as a resource before making changes to it.
If resource fallback to shared triggered, no changes to
module are made. An API is added to finish changing a module
to allow other agents to get resource as shared.

The module is made volatile since configurator rollback for
modules is not required - it is better for modules to stay
loaded after test run. Also module_list() does not list modules
from loaded in the system so that if an agent want to use a
module, it should add it to configurator first.