Bug 11092: add resource acquire fallback
In case of exclusive system-wide resource manipulation (e.g. kernel
modules), an agent may want to make changes to a resource and keep
these changes. A single agent can do that by acquiring a resource
in exclusive mode, making changes and setting the resource to shared
mode, in which all changes are prohibited.

But when several agents want to do the same thing, only one of them will
succeed in acquiring an exclusive resource. To sort out that problem,
a resource fallback is added so that the agents that cannot acquire a
resource in exclusive mode, will acquire the resource in shared mode
after a timeout. Agents can check in which mode a resource was acquired
and if fallback happened, they can rely on the first agent assuming that
it configured the resource correctly and continue to use a resource
in shared mode.