Bug 10896: engine/logger: do not die if no config file is specified
Right now Logger will exit with an error if the path to the config file is
not specified. This behaviour isn't optimal, and it used to be legal not
to specify the config file before. This change in behaviour is the result
of a mistake made during implementation of YAML config parsing.