Bug 11083: engine/tester: assemble test execution plan
An execution plan provides the information needed to understand the total amount
of work (and its structure) that needs to be done during a particular TE run.
This is important for progress tracking and live results.

Now Tester Control messages can reference nodes of this plan in addition to
flow tree nodes.