Bug 11083: reference execution plan items in Tester Control messages
This patch adds a `plan_id` field to MI messages. The value of this field
provides a sequential number of the run item in the execution plan.

The implemented approach handles skips caused by prologue failures. Support
for exceptions is "best-effort", same for the `repeat` attribute.

`skip_start` and `skip_end` may be nested. It's possible to avoid this, but
that would mean more changes to the configuration walking code.