Bug 11245: make module parameters non-volatile
Volatile property is inherited by all configurtor node children, so
the "module" node should be non-volatile to enable automatic restoration
of module parameters. Add a "loaded" node which is made volatile
instead so that the module parameters will be restored on backup, but
the "loaded" state of the module will not change.

Make "filename" property volatile since it is tightly related to the
"loaded" state and cannot be restored out of sync with "loaded".