Bug 11242: engine/builder: pass the path to pkgconfig files via meson option
This patch fixes the build issues on incremental builds from earlier
revisions. The issue was caused by the fact that newer versions of meson
ignore changes in the PKG_CONFIG_PATH variable between builds, so a command
line argument should be used too, since it's never ignored.

This command line option is relatively new, so in some environments a
warning message can be emitted.