Bug 8714: add EAL fini API
The API releases CPUs allocated by the EAL init and restart the
RPC server to completely deinitialize EAL.

The API is required to destroy EAL that was initialized temporarily
to get relevant information in DPDK (e.g. in dpdk-pmd-ts package
prologues).

Make EAL init function always set cached EAL parameters to be able
to release resources even when reuse PCO is not enabled.