cleanup: build: remove partial build options support
Partial build options do not make sense with meson and not supported.
These options are not widely used even in the case of autotools build.
Simplify build system by reducing a number of supported options.

Keep --build-log=* and --build-nolog=* just as a reminder about
feature which makes sense in the case of meson build as well and
we should try to implement it.