Bug 11132: engine/logger: stop test execution on listener's command
This is a useful feature: it allows the user to stop TE if they see that
something already went wrong.

The listener is expected to return a JSON object, with one optional field
called "stop". If the listener's "allow_stop" value is true and the Tester's
PID is known, Tester will be sent a SIGINT. Otherwise, a corresponding error
message will be logged and the execution will proceed as usual.